requirements.txt
boto3
Pillow
pyautogui
keyboard
psutil
cryptography
requests
pytz
